Fixed 64bit issues (int/long)

+++ src/components/general/hashtable.c
@@ -126,9 +126,9 @@ void *ht_find(t_hash_table *ht, const char *key) {
 /**
  * Return value of key, or NULL when nothing found
  */
-void *ht_num_find(t_hash_table *ht, int index) {
+void *ht_num_find(t_hash_table *ht, long index) {
     char key[32];
-    snprintf(key, 31, "%d", index);
+    snprintf(key, 31, "%ld", index);
     return ht_find(ht, key);
 }
 
@@ -146,9 +146,9 @@ int ht_add(t_hash_table *ht, const char *key, void *value) {
     return ht->hashfuncs->add(ht, key, value);
 }
 
-int ht_num_add(t_hash_table *ht, int index, void *value) {
+int ht_num_add(t_hash_table *ht, long index, void *value) {
     char key[32];
-    snprintf(key, 31, "%d", index);
+    snprintf(key, 31, "%ld", index);
     return ht_add(ht, key, value);
 }
